requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.44
psycopg2-binary>=2.9.11
passlib>=1.7.4
bcrypt>=4.1.3
email-validator>=2.3.0
pydantic>=2.8.2
python-multipart>=0.0.9
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
langchain>=0.1.16
langchain-huggingface>=0.0.1
sentence-transformers
chromadb
huggingface-hub
llama-cpp-python
langchain-community
numpy>=2.0.0
qwen-vl-utils
prometheus-fastapi-instrumentator
influxdb-client
pymupdf
google-generativeai>=0.3.0
langchain-google-genai>=1.0.0
requests>=2.31.0
geopy
twilio
langgraph
mcp
duckduckgo-search
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
neurokit2
tsfresh
pandas
scipy
google-search-results
websockets>=12.0
